requirements.txt
Pyrogram==1.4.8
python-dotenv==0.19.2
